introduce TDefDeserialiser interface
This allows to unify the implementation for the Bytes and
JSON Parsers. The String parser cannot be retrofit with
this API since it makes use of a dependent parser. We can
however imply implement the `deserialise` for it.

This also gets rid of an awkward situation when parsing
products in the JSON format where the compiler would
"forget" that a the first projection of a zipped list
is the same as the first list used for the zip. This
requires an additional `ParseProd` type but gets rid
of any ambiguity